working drive mode, needs to be verified
